Add basic scaffolding and logic to verify a URL and fetch data from it
This change adds a small CLI tool, a URL verification utility and a basic
web crawler that takes a single base URL as input and extracts all the URLs
from its webpage.

It also adds problem_statement.md to describe the problem, and adds details
in the README about running the tool, running specs and the outline of the
approach we want to use to implement the crawler.