Add protection against redirect loops
A common problem that web crawlers face is getting stuck in infinite redirect loops. This is when URL A redirects to B and B back to A.
It can also happen with more URLs in a way such as a -> B -> C -> A. We have to specify a max length for a redirect chain. There can
be much longer chains that are difficult to protect against. But with this change there is protection for most common cases.